1.8.3
m_iMaterSel 應為 int (DataCenter.h)

+++ CMDXMoldTesting/DataCenter.h
@@ -62,7 +62,7 @@ public:
 	{
 		m_iMaterialSel = material;
 	}
-	double GetMaterialSel()
+	int GetMaterialSel()
 	{
 		return m_iMaterialSel;
 	}
